+++ server/app.js
@@ -6,6 +6,8 @@ const bannnerRoute = require("./routes/banner-routes");
 const authRoute = require("./routes/auth-routes");
 const productRoute = require("./routes/product-routes");
 const errorMiddleware = require("./middleware/error-middleware");
+const adminProductRoutes = require("./routes/admin-routes/admin-product-routes");
+const adminBannerRoutes = require("./routes/admin-routes/admin-banner-route");
 const cors = require("cors");
 
 // handling cors error
@@ -21,6 +23,9 @@ app.use(express.json());
 app.use("/api", bannnerRoute);
 app.use("/api/auth", authRoute);
 app.use("/api/product", productRoute);
+app.use("/api/admin/product", adminProductRoutes);
+app.use("/api/admin/banner", adminBannerRoutes);
+
 
 app.use(errorMiddleware);
